requirements.txt
numpy
pandas
networkx
scikit-learn
sentence-transformers
umap-learn
matplotlib
pytest
